requirements.txt
pytest
numpy
scikit-learn
sentence-transformers
umap-learn